% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_ACTIVITY_ANZSIC06}
\alias{CL_ACTIVITY_ANZSIC06}
\title{Codelist Activity - ISIC, Revision 4}
\format{
A data frame with 825 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing economic activities according to ANZSIC.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_ACTIVITY_ANZSIC06
}
\description{
A dataset containing the \code{	Activity - ISIC, Revision 4} codelist,
abbreviated as \code{CL_ACTIVITY_ANZSIC06}. Version 1.0.
}
\details{
The SDMX description: "This code list provides lists of values for describing
the economic activity, defined as the combination of actions that result in the
production, distribution and consumption of goods or services. It is based on the
Australian/New Zealand standard classification ANZSIC (Australian and New Zealand
Standard Industrial Classification), 2006."
}
\keyword{datasets}
