% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_ACTIVITY_ISIC4}
\alias{CL_ACTIVITY_ISIC4}
\title{Codelist Activity - ISIC, Revision 4}
\format{
A data frame with 766 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing economic activities.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_ACTIVITY_ISIC4
}
\description{
A dataset containing the \code{Activity - ISIC, Revision 4} codelist,
abbreviated as \code{CL_ACTIVITY_ISIC4}. Version 1.0.
}
\details{
The SDMX description: "This code list provides lists of values for describing
the economic activity, defined as the combination of actions that result in the
production, distribution and consumption of goods or services. It is based on the
United Nations standard classification ISIC (International Standard Industrial
Classification of All Economic Activities), Revision 4 (2008)."
}
\keyword{datasets}
