% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_ACTIVITY_NACE2}
\alias{CL_ACTIVITY_NACE2}
\title{Codelist Activity - NACE, Revision 2}
\format{
A data frame with 996 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing economic activities according to NACE Rev.2.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_ACTIVITY_NACE2
}
\description{
A dataset containing the \code{Activity - NACE, Revision 2} codelist,
abbreviated as \code{CL_ACTIVITY_NACE2}. Version 1.0.
}
\details{
The SDMX description: 'This code list provides lists of values for describing the economic activity, defined as the combination of actions that result in the production, distribution and consumption of goods or services. It is based on the European Union standard classification NACE (Statistical Classification of Economic Activities in the European Community), Revision 2 (2008). More information about and
supporting material for this code list and SDMX code lists in general
(e.g. list of generic codes for expressing general concepts like "Total", "Unknown",
etc.; syntaxes for the creation of further codes; general guidelines for the creation
of SDMX code lists) can be found at this address: \url{https://sdmx.org/?page_id=4345}.'
}
\keyword{datasets}
