% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_AGE}
\alias{CL_AGE}
\title{Codelist Age}
\format{
A data frame with 5 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing age-related time concepts.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_AGE
}
\description{
A dataset containing the \code{Age} codelist,
abbreviated as \code{CL_AGE}. Version 1.0.
}
\details{
The SDMX description: "This code list provides a set of building blocks
to be used for creating simple or complex code identifiers relating to the concept of
\code{age} as defined in the Cross-Domain Concepts and the Metadata Common Vocabulary,
i.e. the length of time that a person has lived or a thing has existed.
This code list was formally adopted on 7 February 2014."
}
\keyword{datasets}
