% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_AREA}
\alias{CL_AREA}
\title{Reference Area Code List}
\format{
A data frame with 899 rows and 5 variables:
\describe{
\item{id}{A statistical code lise for describing reference geographical areas.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code, with reference to the relevant international standard.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_AREA
}
\description{
A dataset contains \code{Reference Area Code List},
abbreviated as \code{CL_AREA}. Version 2.0.
}
\details{
The SDMX description: 'Represents a combination of
reference area codes in the [United Nations] \href{https://unstats.un.org/unsd/methodology/m49/}{M49} and \href{https://www.iso.org/iso-3166-country-codes.html}{ISO-3166} classifications.'
}
\keyword{datasets}
