% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_CONF_STATUS}
\alias{CL_CONF_STATUS}
\title{Codelist Confidentiality Status}
\format{
A data frame with 11 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the confidentiality status of a statistical product.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_CONF_STATUS
}
\description{
A dataset containing the \code{Confidentiality Status} codelist,
abbreviated as \code{CL_CONF_STATUS}. Version 1.2.
}
\details{
Metadata for describing the confidentiality of a statistical product, from
\code{Free (free for publication)} till \code{Not for publication, restricted for internal use only}.
}
\keyword{datasets}
