% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_COPNI_1999}
\alias{CL_COPNI_1999}
\title{Classification of the Purposes of Non-Profit Institutions Serving Households}
\format{
A data frame with 65 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing non-profit services for households.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_COPNI_1999
}
\description{
A dataset containing the \code{Classification of the Purposes of Non-Profit Institutions Serving Households} codelist,
abbreviated as \code{CL_COPNI_1999}. Version 1.0.
}
\keyword{datasets}
