% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_COPP_1999}
\alias{CL_COPP_1999}
\title{Classification of the Outlays of Producers According to Purpose}
\format{
A data frame with 51 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing outlays to producers.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_COPP_1999
}
\description{
A dataset containing the \code{Classification of the Outlays of Producers According to Purpose} codelist,
abbreviated as \code{CL_COPP_1999}. Version 1.0.

SDMX desciption: "This code list provides a set of building blocks to be
used for classifying outlays of enterprises by purpose. It is based on the United Nations
eponymous standard classification. This code list was formally adopted on 27 June 2014."
}
\keyword{datasets}
