% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_DECIMALS}
\alias{CL_DECIMALS}
\title{Codelist Decimals}
\format{
A data frame with 16 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the number of decimals.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_DECIMALS
}
\description{
A dataset containing the \code{Decimals} codelist,
abbreviated as \code{CL_DECIMALS}. Version 1.0.
}
\details{
This code list provides a list of values showing the number of decimal
digits used in the data. This code list was released in 2009.
}
\keyword{datasets}
