% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_DEG_URB}
\alias{CL_DEG_URB}
\title{Codelist Degree of Urbanization}
\format{
A data frame with 14 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing degrees of urbanizations.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_DEG_URB
}
\description{
A dataset containing the \code{Degree of Urbanization} codelist,
abbreviated as \code{CL_DEG_URB}. Version 1.0.
}
\details{
The SDMX description: "This code list provides coded information about
the degree to which area is urbanised."
}
\keyword{datasets}
