% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_OCCUPATION}
\alias{CL_OCCUPATION}
\title{Codelist Occupation}
\format{
A data frame with 619 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing occupations.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_OCCUPATION
}
\description{
A dataset containing the \code{Occupation} codelist,
abbreviated as \code{CL_OCCUPATION}. Version 2.2.
}
\details{
The SDMX description: "This code list provides a list of values for
describing the job or position held by an individual."
}
\keyword{datasets}
