% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_SEASONAL_ADJUST}
\alias{CL_SEASONAL_ADJUST}
\title{Codelist Seasonal Adjustment}
\format{
A data frame with 11 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing seasonal adjustments.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_SEASONAL_ADJUST
}
\description{
Codelist \code{Seasonal Adjustment}, abbreviated \code{CL_SEASONAL_ADJUST} Version 1.0
}
\details{
The SDMX description: 'Seasonal adjustment is a statistical technique used to
remove the effects of seasonal and calendar influences operating on a data series.
Version 1.0 of the code list was formally adopted by the SDMX Secretariat on 4 July 2014.
'
}
\keyword{datasets}
