% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_SEX}
\alias{CL_SEX}
\title{Codelist Sex}
\format{
A data frame with 7 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing sex.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_local}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_SEX
}
\description{
Codelist \code{Sex}, Version 2.1.
}
\details{
The SDMX description: "This code list provides coded information
about the state of being male or female.
This concept is applied if data needs to be categorised by sex.
It applies not only to human beings but also to animals and other living organisms."
}
\keyword{datasets}
