% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_TIME_FORMAT}
\alias{CL_TIME_FORMAT}
\title{Codelist Time Format}
\format{
A data frame with 21 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the various time formats.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_TIME_FORMAT
}
\description{
Codelist \code{Time Format}, abbreviated as \code{CL_TIME_FORMAT}. Version 1.0.
}
\details{
The SDMX description: 'This code list provides coded information (based on the ISO 8601 standard)
indicating the type of time references used in the data. It was released in 2009. More information about this
code list and SDMX code lists in general (e.g. list of generic codes for expressing general concepts like "Total",
"Unknown", etc.; syntaxes for the creation of further codes; general guidelines for the creation of SDMX code
lists) can be found at this address: \url{https://sdmx.org/?page_id=4345}.'
}
\keyword{datasets}
