% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL_TIME_PER_COLLECT.R, R/data.R
\docType{data}
\name{CL_TIME_PER_COLLECT}
\alias{CL_TIME_PER_COLLECT}
\title{Codelist Time Period - Collection}
\format{
A data frame with 47 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the transformation of the time series.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_local}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}

A data frame with 47 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the transformation of the time series.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_local}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}

\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_TIME_PER_COLLECT

CL_TIME_PER_COLLECT
}
\description{
Codelist \code{Time Period - Collection}, abbreviated as \code{CL_TIME_PER_COLLECT}. Version 1.0.

Codelist \code{Time Period - Collection}, abbreviated as \code{CL_TIME_PER_COLLECT}. Version 1.0.
}
\details{
The SDMX description: "	This code list provides a list of values for representing
dates or periods during which the observations have been collected in order to compile
the indicator for the target reference period. This code list was formally adopted on
15 April 2015. More information about this code list and SDMX code lists in general
(e.g. list of generic codes for expressing general concepts like "Total", "Unknown",
etc.; syntaxes for the creation of further codes; general guidelines for the creation of
SDMX code lists) can be found at this address: \url{https://sdmx.org/?page_id=4345}."

The SDMX description: 'This code list provides a list of values for representing
dates or periods during which the observations have been collected in order to compile
the indicator for the target reference period. This code list was formally adopted on
15 April 2015. More information about this code list and SDMX code lists in general
(e.g. list of generic codes for expressing general concepts like "Total", "Unknown",
etc.; syntaxes for the creation of further codes; general guidelines for the creation of
SDMX code lists) can be found at this address: \url{https://sdmx.org/?page_id=4345}.'
}
\keyword{datasets}
