% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_TIMETRANS}
\alias{CL_TIMETRANS}
\title{Codelist Time Transformation}
\format{
A data frame with 47 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the transformation types of time series.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_TIMETRANS
}
\description{
Codelist \code{Time Transformation}, abbreviated as
\code{CL_TIMETRANS}. Version 1.0.
}
\details{
The SDMX description: "Code list for compound coding (type and period)
of transformation types of time series."
}
\keyword{datasets}
