% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_TIMETRANS_PER}
\alias{CL_TIMETRANS_PER}
\title{Codelist Time Transformation Period}
\format{
A data frame with 12 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the transformation periods of time series.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_TIMETRANS_PER
}
\description{
Codelist \code{Time Transformation Period}, abbreviated as
\code{CL_TIMETRANS_PER}. Version 1.0.
}
\details{
The SDMX description: "	Code list for information about the number
of periods used for a time-related transformation of the time series."
}
\keyword{datasets}
