% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_TIMETRANS_TYPE}
\alias{CL_TIMETRANS_TYPE}
\title{Codelist Time Transformation Type}
\format{
A data frame with 13 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the transformation types of time series.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_TIMETRANS_TYPE
}
\description{
Codelist \code{Time Transformation Type}, abbreviated as \code{CL_TIMETRANS_TYPE}. Version 1.0.
}
\details{
The SDMX description: "Code list for time-related transformation types of time series."
}
\keyword{datasets}
