% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_UNIT_MULT}
\alias{CL_UNIT_MULT}
\title{Codelist Unit multiplier}
\format{
A data frame with 31 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing unit multipliers.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_UNIT_MULT
}
\description{
Codelist \code{Unit multiplier}, abbreviated as \code{CL_UNIT_MULT}. Version 1.1.
}
\details{
The SDMX description: "This code list provides code values for indicating the
magnitude in the units of measurement. More information about this code list and SDMX
code lists in general (e.g. list of generic codes for expressing general concepts like
"Total", "Unknown", etc.; syntaxes for the creation of further codes; general guidelines
for the creation of SDMX code lists) can be found at this address:
\url{https://sdmx.org/?page_id=4345}."
}
\keyword{datasets}
