% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_CIVIL_STATUS}
\alias{CL_CIVIL_STATUS}
\title{Codelist Civil (or Marital) Status}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the civil (marital) status of the individual.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_CIVIL_STATUS
}
\description{
A dataset containing the \code{Civil (or Marital) Status} codelist,
abbreviated as \code{CL_CIVIL_STATUS}. Version 1.0.
}
\details{
The SDMX description: "This code list provides a list of values for describing
the civil (or marital) status of an individual, i.e. the legal, conjugal status of an individual in
relation to the marriage laws or customs of the country.
This code list was formally adopted on 25 November 2013."
}
\keyword{datasets}
