% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_OBS_STATUS}
\alias{CL_OBS_STATUS}
\title{Codelist Observation status}
\format{
A data frame with 20 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing data observation status.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_OBS_STATUS
}
\description{
A dataset containing the \code{Observation Status} codelist,
abbreviated as \code{CL_OBS_STATUS}. Version 2.2.
}
\details{
The SDMX description: "This code list provides coded information about the
"status" of an observation (with respect to events such as the ones reflected in the
codes composing the code list). The observation status provides information on
\enumerate{
\item the quality of a value or 2) unusual or missing values."
}
}
\keyword{datasets}
