% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CL_TIME_PER_COLLECT}
\alias{CL_TIME_PER_COLLECT}
\title{Codelist Time Period - Collection}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{id}{A statistical code for describing the collection method within a time frame.}
\item{name}{The standard label of the code.}
\item{description}{The description of the code.}
\item{name_locale}{The language code of the name label.}
\item{description_locale}{The language code of the description.}
}
}
\source{
\url{https://sdmx.org/?page_id=3215/}
}
\usage{
CL_TIME_PER_COLLECT
}
\description{
Codelist \code{Time Period - Collection},
abbreviated as \code{CL_TIME_PER_COLLECT}. Version 1.0.
}
\details{
The SDMX description: "This code list provides a list of values for representing
dates or periods during which the observations have been collected in order to compile
the indicator for the target reference period. This code list was formally adopted on
15 April 2015."
}
\keyword{datasets}
